need to implement bandwith control need to implement choke/unchoke strategy need to add DHT then support magnet links

+++ sicTorrent/src/Torrent.java
@@ -548,6 +548,8 @@ public class Torrent implements Serializable {
     }
 
     public void killThreads() {
+        if (status == TorrentStatus.INACTIVE)
+            return;
         trackermanager.kill();
         peermanager.kill();
         synchronized (peers) {
@@ -578,6 +580,8 @@ public class Torrent implements Serializable {
     }
 
     public void invokeThreads() {
+        if (status == TorrentStatus.ACTIVE)
+            return;
         peers = Collections.synchronizedMap(new HashMap<>());
         peermanager = new PeerManager(50);
         if (trackermanager==null) {trackermanager = new TrackerManager();}
